remove the pydantic_definitions folder and put files in schemas folder
