+++ pymetadataeditor/tools.py
@@ -1,17 +1,9 @@
 import json
 from typing import Dict, Type, Union
 
-from pydantic import BaseModel, ConfigDict
+from pydantic import BaseModel
 
-
-class SchemaBaseModel(BaseModel):
-    model_config = ConfigDict(
-        validate_assignment=True, protected_namespaces=()
-    )  # if a subclass has a model_config then this will be overridden
-
-    def __setitem__(self, key, value):
-        """Allow dict like setting: Model[key] = value"""
-        setattr(self, key, value)
+from pymetadataeditor.schemas.common_schemas import SchemaBaseModel
 
 
 def validate_metadata(metadata: Union[Dict, SchemaBaseModel, str], schema_definition: Type[SchemaBaseModel]) -> None:
