+++ tests/test_tools.py
@@ -2,8 +2,8 @@ import json
 
 import pytest
 
-import pymetadataeditor.schemas.pydantic_definitions.timeseries_schema as tss
-from pymetadataeditor.schemas.pydantic_definitions.common_schemas import SchemaBaseModel
+import pymetadataeditor.schemas.timeseries_schema as tss
+from pymetadataeditor.schemas.common_schemas import SchemaBaseModel
 from pymetadataeditor.tools import update_metadata, validate_metadata
